Move schedule API to its' own files
