+++ internal/internal_schedule_client.go
@@ -1,0 +1,783 @@
+// The MIT License
+//
+// Copyright (c) 2022 Temporal Technologies Inc.  All rights reserved.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package internal
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"github.com/pborman/uuid"
+	commonpb "go.temporal.io/api/common/v1"
+	enumspb "go.temporal.io/api/enums/v1"
+	schedulepb "go.temporal.io/api/schedule/v1"
+	"go.temporal.io/api/serviceerror"
+	taskqueuepb "go.temporal.io/api/taskqueue/v1"
+	workflowpb "go.temporal.io/api/workflow/v1"
+	"go.temporal.io/api/workflowservice/v1"
+	"go.temporal.io/sdk/converter"
+	"go.temporal.io/sdk/internal/common"
+)
+
+type (
+	scheduleHandleImpl struct {
+		ID string
+		client     *WorkflowClient
+	}
+
+	// scheduleListIteratorImpl is the implementation of ScheduleListIterator
+	scheduleListIteratorImpl struct {
+		// whether this iterator is initialized
+		initialized bool
+		// local cached history events and corresponding consuming index
+		nextScheduleIndex int
+		schedules         []*schedulepb.ScheduleListEntry
+		// token to get next page of history events
+		nextToken []byte
+		// err when getting next page of history events
+		err error
+		// func which use a next token to get next page of history events
+		paginate func(nexttoken []byte) (*workflowservice.ListSchedulesResponse, error)
+	}
+)
+
+// Required to implement ClientOutboundInterceptor
+func (*workflowClientInterceptor) mustEmbedClientOutboundInterceptorBase() {}
+
+type scheduleClientInterceptor struct{ client *WorkflowClient }
+
+func (s *scheduleClientInterceptor) CreateSchedule(ctx context.Context, in *ScheduleClientCreateInput) (ScheduleHandle, error) {
+	// This is always set before interceptor is invoked
+	ID := in.Options.ID
+	if ID == "" {
+		return nil, fmt.Errorf("no schedule ID in options")
+	}
+
+	dataConverter := WithContext(ctx, s.client.dataConverter)
+	if dataConverter == nil {
+		dataConverter = converter.GetDefaultDataConverter()
+	}
+
+	action, err := convertToPBScheduleAction(ctx, s.client, in.Options.Action)
+	if err != nil {
+		return nil, err
+	}
+
+	memo, err := getWorkflowMemo(in.Options.Memo, dataConverter)
+	if err != nil {
+		return nil, err
+	}
+
+	searchAttr, err := serializeSearchAttributes(in.Options.SearchAttributes)
+	if err != nil {
+		return nil, err
+	}
+
+	var triggerImmediately *schedulepb.TriggerImmediatelyRequest
+	if in.Options.TriggerImmediately {
+		triggerImmediately = &schedulepb.TriggerImmediatelyRequest{
+			OverlapPolicy: in.Options.Overlap,
+		}
+	}
+
+	backfillRequests := convertToPBBackfillList(in.Options.ScheduleBackfill)
+
+	// run propagators to extract information about tracing and other stuff, store in headers field
+	startRequest := &workflowservice.CreateScheduleRequest{
+		Namespace:  s.client.namespace,
+		ScheduleId: ID,
+		RequestId:  uuid.New(),
+		Schedule: &schedulepb.Schedule{
+			Spec:   convertToPBScheduleSpec(&in.Options.Spec),
+			Action: action,
+			Policies: &schedulepb.SchedulePolicies{
+				OverlapPolicy:  in.Options.Overlap,
+				CatchupWindow:  &in.Options.CatchupWindow,
+				PauseOnFailure: in.Options.PauseOnFailure,
+			},
+			State: &schedulepb.ScheduleState{
+				Notes:            in.Options.Note,
+				Paused:           in.Options.Paused,
+				LimitedActions:   in.Options.RemainingActions != 0,
+				RemainingActions: in.Options.RemainingActions,
+			},
+		},
+		InitialPatch: &schedulepb.SchedulePatch{
+			TriggerImmediately: triggerImmediately,
+			BackfillRequest:    backfillRequests,
+		},
+		Identity:         s.client.identity,
+		Memo:             memo,
+		SearchAttributes: searchAttr,
+	}
+
+	grpcCtx, cancel := newGRPCContext(ctx, defaultGrpcRetryParameters(ctx))
+	defer cancel()
+
+	_, err = s.client.workflowService.CreateSchedule(grpcCtx, startRequest)
+	if _, ok := err.(*serviceerror.WorkflowExecutionAlreadyStarted); ok {
+		return nil, ErrScheduleAlreadyRunning
+	}
+	if err != nil {
+		return nil, err
+	}
+
+	return &scheduleHandleImpl{
+		ID: ID,
+		client:     s.client,
+	}, nil
+}
+
+func (s *scheduleClientInterceptor) mustEmbedScheduleClientInterceptor() {}
+
+func (wc *WorkflowClient) Create(ctx context.Context, options ScheduleOptions) (ScheduleHandle, error) {
+	if err := wc.ensureInitialized(); err != nil {
+		return nil, err
+	}
+
+	// Set header before interceptor run
+	ctx = contextWithNewHeader(ctx)
+
+	// Run via interceptor
+	return wc.scheduleInterceptor.CreateSchedule(ctx, &ScheduleClientCreateInput{
+		Options: &options,
+	})
+}
+
+func (wc *WorkflowClient) GetHandle(ctx context.Context, scheduleID string) ScheduleHandle {
+	return &scheduleHandleImpl{
+		ID: scheduleID,
+		client:     wc,
+	}
+}
+
+func (wc *WorkflowClient) List(ctx context.Context, options ScheduleListOptions) (ScheduleListIterator, error) {
+	paginate := func(nextToken []byte) (*workflowservice.ListSchedulesResponse, error) {
+		grpcCtx, cancel := newGRPCContext(ctx, defaultGrpcRetryParameters(ctx))
+		defer cancel()
+		request := &workflowservice.ListSchedulesRequest{
+			Namespace:       wc.namespace,
+			MaximumPageSize: options.PageSize,
+			NextPageToken:   nextToken,
+		}
+		var response *workflowservice.ListSchedulesResponse
+		var err error
+		for {
+			response, err = wc.workflowService.ListSchedules(grpcCtx, request)
+			if err != nil {
+				return nil, err
+			}
+			if len(response.Schedules) == 0 && len(response.NextPageToken) != 0 {
+				request.NextPageToken = response.NextPageToken
+				continue
+			}
+			break
+		}
+		return response, nil
+	}
+
+	return &scheduleListIteratorImpl{
+		paginate: paginate,
+	}, nil
+}
+
+func (iter *scheduleListIteratorImpl) HasNext() bool {
+	if iter.nextScheduleIndex < len(iter.schedules) || iter.err != nil {
+		return true
+	} else if !iter.initialized || len(iter.nextToken) != 0 {
+		iter.initialized = true
+		response, err := iter.paginate(iter.nextToken)
+		iter.nextScheduleIndex = 0
+		if err == nil {
+			iter.schedules = response.Schedules
+			iter.nextToken = response.NextPageToken
+			iter.err = nil
+		} else {
+			iter.schedules = nil
+			iter.nextToken = nil
+			iter.err = err
+		}
+
+		if iter.nextScheduleIndex < len(iter.schedules) || iter.err != nil {
+			return true
+		}
+		return false
+	}
+
+	return false
+}
+
+func (iter *scheduleListIteratorImpl) Next() (*ScheduleListEntry, error) {
+	// if caller call the Next() when iteration is over, just return nil, nil
+	if !iter.HasNext() {
+		panic("ScheduleListIterator Next() called without checking HasNext()")
+	}
+
+	// we have cached events
+	if iter.nextScheduleIndex < len(iter.schedules) {
+		index := iter.nextScheduleIndex
+		iter.nextScheduleIndex++
+		schedule := iter.schedules[index]
+		return convertFromPBScheduleListEntry(schedule), nil
+	} else if iter.err != nil {
+		// we have err, clear that iter.err and return err
+		err := iter.err
+		iter.err = nil
+		return nil, err
+	}
+	panic("ScheduleListIterator Next() should return either a schedule or a err")
+}
+
+func (scheduleHandle *scheduleHandleImpl) GetID() string {
+	return scheduleHandle.ID
+}
+
+func (scheduleHandle *scheduleHandleImpl) Delete(ctx context.Context) error {
+	request := &workflowservice.DeleteScheduleRequest{
+		Namespace:  scheduleHandle.client.namespace,
+		ScheduleId: scheduleHandle.ID,
+		Identity:   scheduleHandle.client.identity,
+	}
+	grpcCtx, cancel := newGRPCContext(ctx, defaultGrpcRetryParameters(ctx))
+	defer cancel()
+	_, err := scheduleHandle.client.workflowService.DeleteSchedule(grpcCtx, request)
+	return err
+}
+
+func (scheduleHandle *scheduleHandleImpl) Backfill(ctx context.Context, options ScheduleBackfillOptions) error {
+	request := &workflowservice.PatchScheduleRequest{
+		Namespace:  scheduleHandle.client.namespace,
+		ScheduleId: scheduleHandle.ID,
+		Patch: &schedulepb.SchedulePatch{
+			BackfillRequest: convertToPBBackfillList(options.Backfill),
+		},
+		Identity:  scheduleHandle.client.identity,
+		RequestId: uuid.New(),
+	}
+	grpcCtx, cancel := newGRPCContext(ctx, defaultGrpcRetryParameters(ctx))
+	defer cancel()
+	_, err := scheduleHandle.client.workflowService.PatchSchedule(grpcCtx, request)
+	return err
+}
+
+func (scheduleHandle *scheduleHandleImpl) Update(ctx context.Context, update func(*ScheduleDescription) *ScheduleUpdate, options ScheduleUpdateOptions) error {
+	grpcCtx, cancel := newGRPCContext(ctx, defaultGrpcRetryParameters(ctx))
+	defer cancel()
+	ctx = contextWithNewHeader(ctx)
+
+	describeRequest := &workflowservice.DescribeScheduleRequest{
+		Namespace:  scheduleHandle.client.namespace,
+		ScheduleId: scheduleHandle.ID,
+	}
+	describeResponse, err := scheduleHandle.client.workflowService.DescribeSchedule(grpcCtx, describeRequest)
+	if err != nil {
+		return err
+	}
+	scheduleDescription, err := scheduleDescriptionFromPB(describeResponse)
+	if err != nil {
+		return err
+	}
+	newSchedule := update(scheduleDescription)
+	if newSchedule == nil {
+		return nil
+	}
+	newSchedulePB, err := convertToPBSchedule(ctx, scheduleHandle.client, newSchedule.Schedule)
+	if err != nil {
+		return err
+	}
+	_, err = scheduleHandle.client.workflowService.UpdateSchedule(grpcCtx, &workflowservice.UpdateScheduleRequest{
+		Namespace:     scheduleHandle.client.namespace,
+		ScheduleId:    scheduleHandle.ID,
+		Schedule:      newSchedulePB,
+		ConflictToken: nil,
+		Identity:      scheduleHandle.client.identity,
+		RequestId:     uuid.New(),
+	})
+	return err
+}
+
+func (scheduleHandle *scheduleHandleImpl) Describe(ctx context.Context) (*ScheduleDescription, error) {
+	request := &workflowservice.DescribeScheduleRequest{
+		Namespace:  scheduleHandle.client.namespace,
+		ScheduleId: scheduleHandle.ID,
+	}
+	grpcCtx, cancel := newGRPCContext(ctx, defaultGrpcRetryParameters(ctx))
+	defer cancel()
+	describeResponse, err := scheduleHandle.client.workflowService.DescribeSchedule(grpcCtx, request)
+	if err != nil {
+		return nil, err
+	}
+	return scheduleDescriptionFromPB(describeResponse)
+}
+
+func (scheduleHandle *scheduleHandleImpl) Trigger(ctx context.Context, options ScheduleTriggerOptions) error {
+	request := &workflowservice.PatchScheduleRequest{
+		Namespace:  scheduleHandle.client.namespace,
+		ScheduleId: scheduleHandle.ID,
+		Patch: &schedulepb.SchedulePatch{
+			TriggerImmediately: &schedulepb.TriggerImmediatelyRequest{
+				OverlapPolicy: options.Overlap,
+			},
+		},
+		Identity:  scheduleHandle.client.identity,
+		RequestId: uuid.New(),
+	}
+	grpcCtx, cancel := newGRPCContext(ctx, defaultGrpcRetryParameters(ctx))
+	defer cancel()
+	_, err := scheduleHandle.client.workflowService.PatchSchedule(grpcCtx, request)
+	return err
+}
+
+func (scheduleHandle *scheduleHandleImpl) Pause(ctx context.Context, options SchedulePauseOptions) error {
+	pauseNote := "Paused via Go SDK"
+	if options.Note != "" {
+		pauseNote = options.Note
+	}
+	request := &workflowservice.PatchScheduleRequest{
+		Namespace:  scheduleHandle.client.namespace,
+		ScheduleId: scheduleHandle.ID,
+		Patch: &schedulepb.SchedulePatch{
+			Pause: pauseNote,
+		},
+		Identity:  scheduleHandle.client.identity,
+		RequestId: uuid.New(),
+	}
+	grpcCtx, cancel := newGRPCContext(ctx, defaultGrpcRetryParameters(ctx))
+	defer cancel()
+	_, err := scheduleHandle.client.workflowService.PatchSchedule(grpcCtx, request)
+	return err
+}
+
+func (scheduleHandle *scheduleHandleImpl) Unpause(ctx context.Context, options ScheduleUnpauseOptions) error {
+	unpauseNote := "Unpaused via Go SDK"
+	if options.Note != "" {
+		unpauseNote = options.Note
+	}
+	request := &workflowservice.PatchScheduleRequest{
+		Namespace:  scheduleHandle.client.namespace,
+		ScheduleId: scheduleHandle.ID,
+		Patch: &schedulepb.SchedulePatch{
+			Unpause: unpauseNote,
+		},
+		Identity:  scheduleHandle.client.identity,
+		RequestId: uuid.New(),
+	}
+	grpcCtx, cancel := newGRPCContext(ctx, defaultGrpcRetryParameters(ctx))
+	defer cancel()
+	_, err := scheduleHandle.client.workflowService.PatchSchedule(grpcCtx, request)
+	return err
+}
+
+func convertToPBScheduleSpec(scheduleSpec *ScheduleSpec) *schedulepb.ScheduleSpec {
+	if scheduleSpec == nil {
+		return nil
+	}
+
+	calendar := convertToPBScheduleCalendarSpecList(scheduleSpec.Calendars)
+
+	intervals := make([]*schedulepb.IntervalSpec, len(scheduleSpec.Intervals))
+	for i, interval := range scheduleSpec.Intervals {
+		intervalSpec := interval
+		intervals[i] = &schedulepb.IntervalSpec{
+			Interval: &intervalSpec.Every,
+			Phase:    &intervalSpec.Offset,
+		}
+	}
+
+	skip := convertToPBScheduleCalendarSpecList(scheduleSpec.Skip)
+
+	var startTime *time.Time
+	if !scheduleSpec.StartAt.IsZero() {
+		startTime = &scheduleSpec.StartAt
+	}
+
+	var endTime *time.Time
+	if !scheduleSpec.EndAt.IsZero(){
+		endTime = &scheduleSpec.EndAt
+	}
+
+	return &schedulepb.ScheduleSpec{
+		StructuredCalendar:        calendar,
+		Interval:                  intervals,
+		CronString:                scheduleSpec.CronExpressions,
+		ExcludeStructuredCalendar: skip,
+		StartTime:                 startTime,
+		EndTime:                   endTime,
+		Jitter:                    &scheduleSpec.Jitter,
+		// TODO support custom time zone data
+		TimezoneName: scheduleSpec.TimeZoneName,
+	}
+}
+
+func convertFromPBScheduleSpec(scheduleSpec *schedulepb.ScheduleSpec) *ScheduleSpec {
+	if scheduleSpec == nil {
+		return nil
+	}
+	
+	calendars := convertFromPBScheduleCalendarSpecList(scheduleSpec.GetStructuredCalendar())
+
+	intervals := make([]ScheduleIntervalSpec, len(scheduleSpec.GetInterval()))
+	for i, s := range scheduleSpec.GetInterval() {
+		intervals[i] = ScheduleIntervalSpec{
+			Every:  common.DurationValue(s.Interval),
+			Offset: common.DurationValue(s.Phase),
+		}
+	}
+
+	skip := convertFromPBScheduleCalendarSpecList(scheduleSpec.GetExcludeStructuredCalendar())
+
+	startAt := time.Time{}
+	if scheduleSpec.GetStartTime() != nil {
+		startAt = *scheduleSpec.GetStartTime()
+	}
+
+	endAt := time.Time{}
+	if scheduleSpec.GetEndTime() != nil {
+		endAt = *scheduleSpec.GetEndTime()
+	}
+
+	return &ScheduleSpec{
+		Calendars: calendars,
+		Intervals: intervals,
+		Skip:      skip,
+		StartAt:   startAt,
+		EndAt:     endAt,
+		Jitter:    common.DurationValue(scheduleSpec.GetJitter()),
+		TimeZoneName:  scheduleSpec.GetTimezoneName(),
+	}
+}
+
+func scheduleDescriptionFromPB(describeResponse *workflowservice.DescribeScheduleResponse) (*ScheduleDescription, error) {
+	if describeResponse == nil {
+		return nil, nil
+	}
+
+	runningWorkflows := make([]ScheduleWorkflowExecution, len(describeResponse.Info.GetRunningWorkflows()))
+	for i, s := range describeResponse.Info.GetRunningWorkflows() {
+		runningWorkflows[i] = ScheduleWorkflowExecution{
+			WorkflowID:          s.GetWorkflowId(),
+			FirstExecutionRunID: s.GetRunId(),
+		}
+	}
+
+	recentActions := convertFromPBScheduleActionResultList(describeResponse.Info.GetRecentActions())
+
+	nextActionTimes := make([]time.Time, len(describeResponse.Info.GetFutureActionTimes()))
+	for i, t := range describeResponse.Info.GetFutureActionTimes() {
+		nextActionTimes[i] = common.TimeValue(t)
+	}
+
+	actionDescription, err := convertFromPBScheduleAction(describeResponse.Schedule.Action)
+	if err != nil {
+		return nil, err
+	}
+
+	return &ScheduleDescription{
+		Schedule: Schedule{
+			Action: actionDescription,
+			Spec:   convertFromPBScheduleSpec(describeResponse.Schedule.Spec),
+			Policy: &SchedulePolicies{
+				Overlap:        describeResponse.Schedule.Policies.GetOverlapPolicy(),
+				CatchupWindow:  common.DurationValue(describeResponse.Schedule.Policies.GetCatchupWindow()),
+				PauseOnFailure: describeResponse.Schedule.Policies.GetPauseOnFailure(),
+			},
+			State: &ScheduleState{
+				Note:             describeResponse.Schedule.State.GetNotes(),
+				Paused:           describeResponse.Schedule.State.GetPaused(),
+				LimitedActions:   describeResponse.Schedule.State.GetLimitedActions(),
+				RemainingActions: describeResponse.Schedule.State.GetRemainingActions(),
+			},
+		},
+		Info: ScheduleInfo{
+			NumActions:                    describeResponse.Info.ActionCount,
+			NumActionsMissedCatchupWindow: describeResponse.Info.MissedCatchupWindow,
+			NumActionsSkippedOverlap:      describeResponse.Info.OverlapSkipped,
+			RunningWorkflows:              runningWorkflows,
+			RecentActions:                 recentActions,
+			NextActionTimes:               nextActionTimes,
+			CreatedAt:                     common.TimeValue(describeResponse.Info.GetCreateTime()),
+			LastUpdateAt:                  common.TimeValue(describeResponse.Info.GetUpdateTime()),
+		},
+		Memo:             describeResponse.Memo,
+		SearchAttributes: describeResponse.SearchAttributes,
+	}, nil
+}
+
+func convertToPBSchedule(ctx context.Context, client *WorkflowClient, schedule *Schedule) (*schedulepb.Schedule, error) {
+	if schedule == nil {
+		return nil, nil
+	}
+	action, err := convertToPBScheduleAction(ctx, client, schedule.Action)
+	if err != nil {
+		return nil, err
+	}
+	return &schedulepb.Schedule{
+		Spec:   convertToPBScheduleSpec(schedule.Spec),
+		Action: action,
+		Policies: &schedulepb.SchedulePolicies{
+			OverlapPolicy:  schedule.Policy.Overlap,
+			CatchupWindow:  &schedule.Policy.CatchupWindow,
+			PauseOnFailure: schedule.Policy.PauseOnFailure,
+		},
+		State: &schedulepb.ScheduleState{
+			Notes:            schedule.State.Note,
+			Paused:           schedule.State.Paused,
+			LimitedActions:   schedule.State.LimitedActions,
+			RemainingActions: schedule.State.RemainingActions,
+		},
+	}, nil
+}
+
+func convertFromPBScheduleListEntry(schedule *schedulepb.ScheduleListEntry) *ScheduleListEntry {
+	scheduleInfo := schedule.GetInfo()
+	
+	recentActions := convertFromPBScheduleActionResultList(scheduleInfo.GetRecentActions())
+
+	nextActionTimes := make([]time.Time, len(schedule.Info.GetFutureActionTimes()))
+	for i, t := range schedule.Info.GetFutureActionTimes() {
+		nextActionTimes[i] = common.TimeValue(t)
+	}
+
+	return &ScheduleListEntry{
+		ID: schedule.ScheduleId,
+		Spec:       convertFromPBScheduleSpec(scheduleInfo.GetSpec()),
+		Note:       scheduleInfo.GetNotes(),
+		Paused:     scheduleInfo.GetPaused(),
+		WorkflowType: WorkflowType{
+			Name: scheduleInfo.GetWorkflowType().GetName(),
+		},
+		RecentActions:    recentActions,
+		NextActionTimes:  nextActionTimes,
+		Memo:             schedule.Memo,
+		SearchAttributes: schedule.SearchAttributes,
+	}
+}
+
+func convertToPBScheduleAction(ctx context.Context, client *WorkflowClient, scheduleAction ScheduleAction) (*schedulepb.ScheduleAction, error) {
+	switch action := scheduleAction.(type) {
+	case ScheduleWorkflowAction:
+		// Set header before interceptor run
+		dataConverter := WithContext(ctx, client.dataConverter)
+		if dataConverter == nil {
+			dataConverter = converter.GetDefaultDataConverter()
+		}
+
+		// Default workflow ID
+		if action.WorkflowOptions.ID == "" {
+			action.WorkflowOptions.ID = uuid.New()
+		}
+
+		// Validate function and get name
+		if err := validateFunctionArgs(action.Workflow, action.Args, true); err != nil {
+			return nil, err
+		}
+		workflowType, err := getWorkflowFunctionName(client.registry, action.Workflow)
+		if err != nil {
+			return nil, err
+		}
+		// Encode input
+		input, err := encodeArgs(dataConverter, action.Args)
+		if err != nil {
+			return nil, err
+		}
+		memo, err := getWorkflowMemo(action.WorkflowOptions.Memo, dataConverter)
+		if err != nil {
+			return nil, err
+		}
+
+		searchAttr, err := serializeSearchAttributes(action.WorkflowOptions.SearchAttributes)
+		if err != nil {
+			return nil, err
+		}
+
+		// get workflow headers from the context
+		header, err := headerPropagated(ctx, client.contextPropagators)
+		if err != nil {
+			return nil, err
+		}
+
+		return &schedulepb.ScheduleAction{
+			Action: &schedulepb.ScheduleAction_StartWorkflow{
+				StartWorkflow: &workflowpb.NewWorkflowExecutionInfo{
+					WorkflowId:               action.WorkflowOptions.ID,
+					WorkflowType:             &commonpb.WorkflowType{Name: workflowType},
+					TaskQueue:                &taskqueuepb.TaskQueue{Name: action.WorkflowOptions.TaskQueue, Kind: enumspb.TASK_QUEUE_KIND_NORMAL},
+					Input:                    input,
+					WorkflowExecutionTimeout: &action.WorkflowOptions.WorkflowExecutionTimeout,
+					WorkflowRunTimeout:       &action.WorkflowOptions.WorkflowRunTimeout,
+					WorkflowTaskTimeout:      &action.WorkflowOptions.WorkflowTaskTimeout,
+					WorkflowIdReusePolicy:    action.WorkflowOptions.WorkflowIDReusePolicy,
+					RetryPolicy:              convertToPBRetryPolicy(action.WorkflowOptions.RetryPolicy),
+					Memo:                     memo,
+					SearchAttributes:         searchAttr,
+					Header:                   header,
+				},
+			},
+		}, nil
+	case ScheduleWorkflowActionDescription:
+		// get workflow headers from the context
+		header, err := headerPropagated(ctx, client.contextPropagators)
+		if err != nil {
+			return nil, err
+		}
+		return &schedulepb.ScheduleAction{
+			Action: &schedulepb.ScheduleAction_StartWorkflow{
+				StartWorkflow: &workflowpb.NewWorkflowExecutionInfo{
+					WorkflowId:               action.WorkflowID,
+					WorkflowType:             &commonpb.WorkflowType{Name: action.WorkflowType.Name},
+					TaskQueue:                &taskqueuepb.TaskQueue{Name: action.TaskQueue, Kind: enumspb.TASK_QUEUE_KIND_NORMAL},
+					Input:                    action.Args,
+					WorkflowExecutionTimeout: &action.WorkflowExecutionTimeout,
+					WorkflowRunTimeout:       &action.WorkflowRunTimeout,
+					WorkflowTaskTimeout:      &action.WorkflowTaskTimeout,
+					WorkflowIdReusePolicy:    action.WorkflowIDReusePolicy,
+					RetryPolicy:              convertToPBRetryPolicy(action.RetryPolicy),
+					Memo:                     action.Memo,
+					SearchAttributes:         action.SearchAttributes,
+					Header:                   header,
+				},
+			},
+		}, nil
+	default:
+		// TODO maybe just panic instead?
+		return nil, fmt.Errorf("could not parse ScheduleAction")
+	}
+}
+
+func convertFromPBScheduleAction(action *schedulepb.ScheduleAction) (ScheduleAction, error) {
+	switch action := action.Action.(type) {
+	case *schedulepb.ScheduleAction_StartWorkflow:
+		workflow := action.StartWorkflow
+		return ScheduleWorkflowActionDescription{
+			WorkflowID: workflow.GetWorkflowId(),
+			WorkflowType: WorkflowType{
+				Name: workflow.WorkflowType.GetName(),
+			},
+			Args:                     workflow.Input,
+			TaskQueue:                workflow.TaskQueue.GetName(),
+			WorkflowExecutionTimeout: common.DurationValue(workflow.GetWorkflowExecutionTimeout()),
+			WorkflowRunTimeout:       common.DurationValue(workflow.GetWorkflowRunTimeout()),
+			WorkflowTaskTimeout:      common.DurationValue(workflow.GetWorkflowTaskTimeout()),
+			WorkflowIDReusePolicy:    workflow.GetWorkflowIdReusePolicy(),
+			RetryPolicy:              convertFromPBRetryPolicy(workflow.RetryPolicy),
+			Memo:                     workflow.Memo,
+			SearchAttributes:         workflow.SearchAttributes,
+		}, nil
+	default:
+		// TODO maybe just panic instead?
+		return nil, fmt.Errorf("could not parse ScheduleAction")
+	}
+}
+
+func convertToPBBackfillList(backfillRequests []ScheduleBackfill) []*schedulepb.BackfillRequest {
+	backfillRequestsPB := make([]*schedulepb.BackfillRequest, len(backfillRequests))
+	for i, b := range backfillRequests {
+		backfill := b
+		backfillRequestsPB[i] = &schedulepb.BackfillRequest{
+			StartTime:     &backfill.Start,
+			EndTime:       &backfill.End,
+			OverlapPolicy: backfill.Overlap,
+		}
+	}
+	return backfillRequestsPB
+}
+
+func convertToPBRangeList(scheduleRange []ScheduleRange) []*schedulepb.Range {
+	rangesPB := make([]*schedulepb.Range, len(scheduleRange))
+	for i, r := range scheduleRange {
+		 rangesPB[i] = &schedulepb.Range{
+			Start: int32(r.Start),
+			End:   int32(r.End),
+			Step:  int32(r.Step),
+		}
+	}
+	return rangesPB
+}
+
+func convertFromPBRangeList(scheduleRangePB  []*schedulepb.Range) []ScheduleRange {
+	scheduleRange := make([]ScheduleRange, len(scheduleRangePB))
+	for i, r := range scheduleRangePB {
+		if r == nil {
+			continue
+		}
+		scheduleRange[i] = ScheduleRange{
+			Start: int(r.Start),
+			End:   int(r.End),
+			Step:  int(r.Step),
+		}
+	}
+	return scheduleRange
+}
+
+func convertFromPBScheduleCalendarSpecList(calendarSpecPB []*schedulepb.StructuredCalendarSpec) []ScheduleCalendarSpec {
+	calendarSpec := make([]ScheduleCalendarSpec, len(calendarSpecPB))
+	for i, e := range calendarSpecPB {
+		calendarSpec[i] = ScheduleCalendarSpec{
+			Second:     convertFromPBRangeList(e.Second),
+			Minute:     convertFromPBRangeList(e.Minute),
+			Hour:       convertFromPBRangeList(e.Hour),
+			DayOfMonth: convertFromPBRangeList(e.DayOfMonth),
+			Month:      convertFromPBRangeList(e.Month),
+			Year:       convertFromPBRangeList(e.Year),
+			DayOfWeek:  convertFromPBRangeList(e.DayOfWeek),
+			Comment:    e.Comment,
+		}
+	}
+	return calendarSpec
+}
+
+func convertToPBScheduleCalendarSpecList(calendarSpec []ScheduleCalendarSpec) []*schedulepb.StructuredCalendarSpec{
+	calendarSpecPB := make([]*schedulepb.StructuredCalendarSpec, len(calendarSpec))
+	for i, e := range calendarSpec {
+		calendarSpecPB[i] = &schedulepb.StructuredCalendarSpec{
+			Second:     convertToPBRangeList(e.Second),
+			Minute:     convertToPBRangeList(e.Minute),
+			Hour:       convertToPBRangeList(e.Hour),
+			DayOfMonth: convertToPBRangeList(e.DayOfMonth),
+			Month:      convertToPBRangeList(e.Month),
+			Year:       convertToPBRangeList(e.Year),
+			DayOfWeek:  convertToPBRangeList(e.DayOfWeek),
+			Comment:    e.Comment,
+		}
+	}
+	return calendarSpecPB
+}
+
+func convertFromPBScheduleActionResultList(aa []*schedulepb.ScheduleActionResult) []ScheduleActionResult {
+	recentActions := make([]ScheduleActionResult, len(aa))
+	for i, a := range aa {
+		var workflowExecution *ScheduleWorkflowExecution
+		if a.GetStartWorkflowResult() != nil {
+			workflowExecution = &ScheduleWorkflowExecution{
+				WorkflowID:          a.GetStartWorkflowResult().GetWorkflowId(),
+				FirstExecutionRunID: a.GetStartWorkflowResult().GetRunId(),
+			}
+		}
+		recentActions[i] = ScheduleActionResult{
+			ScheduleTime: common.TimeValue(a.GetScheduleTime()),
+			ActualTime:   common.TimeValue(a.GetActualTime()),
+			StartWorkflowResult: workflowExecution,
+		}
+	}
+	return recentActions
+}
